Refactor reducers into separate files
